pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "verinode"
version = "0.1.0"
description = "Local-first research verification workbench."
readme = "ARCHITECTURE.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "openai>=1.108.0,<2.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
  "integration: tests that exercise external providers",
  "live: tests that use real network calls and credentials",
]
